+++ frysk-sys/lib/dwfl/DwarfDie.java
@@ -261,7 +261,7 @@ abstract public class DwarfDie {
     /**
      * @return The offset for this die.
      */
-    public int getOffset() {
+    public long getOffset() {
 	return get_offset(this.getPointer());
     }
 
@@ -436,7 +436,7 @@ abstract public class DwarfDie {
   
     private native int get_attr_constant (long addr, int attr);
   
-    private native int get_offset (long addr);
+    private native long get_offset (long addr);
 
     // Package access for DwarfDieFactory
     static native int get_tag (long var_die);
